Improved HTML page header title and navtrail info.
When generating HTML page header title information, print either the search
pattern or the collection name or the title of the given record first,
and only then print the server name.  Moreover, do not print record IDs
explicitly in titles/navtrails.  This improves the ranking of the detailed
record pages in general search engines (e.g. Google) when people find words
appearing in titles.  Thanks to Gregory Favre for suggestion.